requirements.txt
discord.py~=2.1
mysql-connector-python~=8.0
Pillow~=9.4
pydantic~=1.10
